requirements.txt
streamlit==1.49.1
ultralytics==8.3.25
torch==2.4.1
torchvision==0.19.1
opencv-python-headless==4.10.0.84
pillow>=10.3
requests>=2.31
pandas==2.3.2